refactor(trusty-common): split memory_client into submodules under 500-SLOC cap (#607)
Split the 754-SLOC memory_client.rs monolith into four focused submodules:
- types.rs (~200 SLOC): constants, URL helpers, wire shapes, public domain types
- parsers.rs (~130 SLOC): all serde_json::Value → domain projections
- client.rs (~150 SLOC): MemoryClient struct and all async HTTP methods
- tests.rs (~270 SLOC): full unit test suite (test-file cap, not prod cap)

The mod.rs re-export facade is ~30 lines. Public API is unchanged.
Removed the old flat-file entry from .line-cap-allowlist.tsv (ratcheted down).
